selective enable outputs plugin using PageConfig
